fix(config): # rename account_id to workspace_id
